fix: address code review issues (architecture + auth + safety)
- Extract SessionFolderQueryPort to application/ports/ (was in use-case
  layer, creating inverted dependency violating Clean Architecture)
- Remove unsafe `as` type cast in container.ts; use separate typed
  variable for DrizzleSessionRepository
- Migrate folders/[id] and folders/reorder routes to withApiAuth
  (mobile API key auth was broken for PATCH/DELETE/reorder)
- Add mounted guards to all setState calls in login catch blocks
- Emit ConnectionStatus.disconnected before closing streams in dispose()
